wip: offers management
