+++ hapi/src/api/index.js
@@ -7,7 +7,6 @@ const userApi = require('./user.api')
 const vaultApi = require('./vault.api')
 const locationApi = require('./location.api')
 const notificationApi = require('./notification.api')
-const offerApi = require('./offer.api')
 
 module.exports = {
   accountApi,
@@ -18,6 +17,5 @@ module.exports = {
   userApi,
   vaultApi,
   locationApi,
-  notificationApi,
-  offerApi
+  notificationApi
 }
