--- hapi/src/routes/offer/offer.route.js
@@ -1,10 +1,0 @@
-const offerHandler = require('./offer.handler')
-
-module.exports = {
-  method: 'POST',
-  path: '/api/offer',
-  handler: offerHandler
-  // options: {
-  //   auth: true
-  // }
-}
